+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Downgrade diagnostic severity level of undefined steps to warnings
+
 ### Added
 - Diagnostics for marking steps as undefined in scenario outlines ([#210](https://github.com/cucumber/language-service/pull/210))
 - Malaysian localisation and translations for "Rule" in Vietnamese, Irish, Danish, Dutch ([#123](https://github.com/cucumber/monaco/pull/123))